Move requests to their own modules
