+++ src/raw/mod.rs
@@ -15,6 +15,7 @@ pub use self::client::Client;
 use std::fmt;
 
 mod client;
+mod requests;
 
 /// A [`ColumnFamily`](ColumnFamily) is an optional parameter for [`raw::Client`](Client) requests.
 ///
